Fix PR remarks:  - cleanup  - move DateTimeOffset parser tests to its own file
